requirements.txt
# Core dependencies
requests>=2.31.0
rich>=13.7.0
Pillow>=10.2.0
tenacity>=8.2.3

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# CLI enhancement
click>=8.1.7
colorama>=0.4.6

# Configuration management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

# Machine Learning & NLP
scikit-learn>=1.3.2
numpy>=1.24.0
scipy>=1.11.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL
pymongo>=4.6.0  # MongoDB (optional)

# Caching
redis>=5.0.0
hiredis>=2.2.3  # Redis performance boost

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI/ML Advanced
google-generativeai>=0.3.0  # Google Gemini AI (FREE)
sentence-transformers>=2.2.2  # Sentence-BERT embeddings
transformers>=4.35.0  # Hugging Face models

# Async & Performance
aiohttp>=3.9.0  # Async HTTP
aiocache>=0.12.2  # Async caching
celery>=5.3.4  # Background tasks
flower>=2.0.1  # Celery monitoring

# Search
elasticsearch>=8.11.0  # Full-text search

# API Management
strawberry-graphql[fastapi]>=0.216.0  # GraphQL
pyotp>=2.9.0  # MFA/TOTP
qrcode>=7.4.2  # QR code generation

# Security
cryptography>=41.0.7  # Encryption
python-multipart>=0.0.6  # Form data parsing

# Testing & Quality Tools
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
responses>=0.24.1
httpx>=0.25.0
coverage>=7.3.0

# Code Quality & Linting
black>=23.12.1
ruff>=0.1.9
flake8>=7.0.0
mypy>=1.8.0
pylint>=3.0.3
isort>=5.13.2
bandit>=1.7.6  # Security linting
safety>=2.3.5  # Dependency security check

# Pre-commit hooks
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.39.0

# Development tools
ipython>=8.18.0
ipdb>=0.13.13
